Increment sequence_number after failure
After an sendSMS submission returns a legitimate exception (i.e. `$response->status != SMPP::ESME_ROK`), the sequence_number isn't incremented. This casues the SMSC to fail all subsequent messages as the sequence number is the same as the original failed submission.

The incrementing of sequence_number on line 624 of smppclient.class.php should be moved prior to the exception on line 622.

If the class supported resubmission attempts this sequence_number should probably not be incremented however as this feature isn't supported, we should just increment it and move on.